requirements.txt
Django==1.6.4
South==0.8.4
decorator==3.4.0
dj-database-url==0.3.0
dj-static==0.0.5
django-bootstrap3==4.4.1
django-countries==2.1.2
django-suit==0.2.8
django-taggit==0.12
django-toolbelt==0.0.1
gunicorn==18.0
lucene-querybuilder==0.2
neo4jrestclient==2.0.2
oauthlib==0.6.1
psycopg2==2.5.2
py2neo==1.6.4
pystache==0.5.3
python-dateutil==2.2
python-openid==2.2.5
python-social-auth==0.1.23
requests==2.2.1
requests-oauthlib==0.4.0
six==1.6.1
static==1.0.2
wsgiref==0.1.2
